Update lineage_clustering branch with changes to master to simplify pull request.

+++ PopPUNK/prune_db.py
@@ -63,7 +63,7 @@ def prune_distance_matrix(refList, remove_seqs_in, distMat, output):
         sys.stderr.write("Removing " + str(len(remove_seqs)) + " sequences\n")
 
         numNew = len(refList) - len(remove_seqs)
-        newDistMat = np.zeros((int(0.5 * numNew * (numNew - 1)), 2))
+        newDistMat = np.zeros((int(0.5 * numNew * (numNew - 1)), 2), dtype=distMat.dtype)
 
         # Create new reference list iterator
         removal_indices.sort()
